+++ plugins/getInterpolator.js
@@ -1,5 +1,3 @@
-/* globals performance */
-
 // The `performance` global is optional
 
 /**
@@ -20,18 +18,17 @@
 function generateUUID () { //  Adapted from original: public domain/MIT: http://stackoverflow.com/a/8809472/271577
   let d = Date.now() +
   // use high-precision timer if available
-  /* eslint-disable compat/compat */
   /* c8 ignore next 4 */
   (typeof performance !== 'undefined' && typeof performance.now === 'function'
     ? performance.now()
-    /* eslint-enable compat/compat */
     : 0);
 
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/gu, function (c) {
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replaceAll(/[xy]/gu, function (c) {
     /* eslint-disable no-bitwise */
-    const r = Math.trunc((d + Math.random() * 16) % 16);
+    // eslint-disable-next-line sonarjs/pseudo-random -- Ok
+    const r = Math.trunc((d + (Math.random() * 16)) % 16);
     d = Math.floor(d / 16);
-    return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
+    return (c === 'x' ? r : ((r & 0x3) | 0x8)).toString(16);
     /* eslint-enable no-bitwise */
   });
 }
@@ -64,7 +61,10 @@ function getInterpolator () {
     },
     plugin: {
       name,
-      set ({element, attribute: {value}, opts}) {
+      set ({
+        // element,
+        attribute: {value}, opts
+      }) {
         // Todo: Support in element name or fragment position
         //  (with ability to convert whole set of arguments?); and
         //  ability to alter all attributes