+++ src/jml-jsdom.js
@@ -1,6 +1,5 @@
 /* eslint-env node */
 
-// eslint-disable-next-line n/no-unpublished-import
 import jsdom from 'jsdom';
 
 import {jml} from './jml.js';