+++ test/chaiESMLoader.js
@@ -1,4 +1,5 @@
-// eslint-disable-next-line no-shadow -- Needed
+// @ts-expect-error Ok
+// eslint-disable-next-line no-undef -- Ok
 const {assert} = chai;
 
 export {