+++ plugins/getInterpolator.mjs
@@ -1,0 +1,3 @@
+import getInterpolator from './getInterpolator.js';
+
+export default getInterpolator;